test: centralize RPC test utilities by creating a dedicated test_utils module
